User checks and specific user endpoint access. For eg. a user who creates a review can only update THAT review

+++ server/Controllers/ReviewController.cs
@@ -1,6 +1,8 @@
 using System.ComponentModel;
 using System.Diagnostics;
+using System.Security.Claims;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using RateMyCollegeClub.Data;
@@ -35,18 +37,37 @@ public class ReviewController : ControllerBase {
     }
 
     [HttpPost]
+    [Authorize]
     public async Task<ActionResult<CreateReviewDTO>> CreateReview(CreateReviewDTO createReviewDTO) {
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        Console.WriteLine("-----------------------------------------------------");
+        Console.WriteLine("HERE IS THE USER ID");
+        Console.WriteLine(userId);
+        if (userId == null) {
+            return Unauthorized();
+        }
+
         var review = _mapper.Map<Review>(createReviewDTO);
+        review.UserId = userId;
+
         await _reviewsRepository.AddAsync(review);
         return CreatedAtAction("GetReview", new {id = review.Id}, review);
     }
 
     [HttpDelete("{id}")]
+    [Authorize]
     public async Task<IActionResult> DeleteReview(int id){
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if(userId == null){
+            return Unauthorized();
+        }
         var review = await _reviewsRepository.GetAsync(id);
 
         if(review is null){
             return NotFound();
+        }
+        if(review.UserId != userId){
+            return Forbid();
         }
 
         await _reviewsRepository.DeleteAsync(id);
@@ -55,11 +76,19 @@ public class ReviewController : ControllerBase {
     }
 
     [HttpPut("{id}")]
+    [Authorize]
     public async Task<IActionResult> UpdateReview(int id, UpdateReviewDTO updateReviewDTO){
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if(userId == null){
+            return Unauthorized();
+        }
         var review = await _reviewsRepository.GetAsync(id);
 
         if(review is null){
             return NotFound();
+        }
+        if(review.UserId != userId){
+            return Forbid();
         }
 
         _mapper.Map(updateReviewDTO, review);
